[modifed] dot application logic
git-svn-id: http://svn2.assembla.com/svn/swgemu/core3/trunk@3713 c3d1530f-68f5-4bd0-87dc-8ef779617e40

+++ MMOCoreORB/src/server/zone/objects/creature/commands/effect/StateEffect.h
@@ -66,7 +66,7 @@ public:
 	virtual ~StateEffect() {}
 
 	void loadState(LuaObject state) {
-		effectType = state.getLongField("effectType");
+		effectType = state.getByteField("effectType");
 		stateChance = state.getIntField("stateChance");
 		stateStrength = state.getIntField("stateStrength");
 		stateLength = state.getIntField("stateLength");
